feat: Model Application
Listed three supervised learning models that are appropriate for this
problem. For each model chosen
Described one real-world application in industry where the model can be
applied.
What are the strengths of the model; when does it perform well?
What are the weaknesses of the model; when does it perform poorly?
What makes this model a good candidate for the problem, given what you
know about the data?